+++ app/Http/Controllers/Admin/SubkegLain/SubkegLainSPPDController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin\SubkegLain;
 
 use App\Http\Controllers\Controller;
 use App\Models\Pegawai;
+use App\Models\RincianBiayaSPPD;
 use App\Models\SPPD;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Http\Request;
@@ -59,7 +60,6 @@ class SubkegLainSPPDController extends Controller
      */
     public function store(Request $request)
     {
-        //
         $this->validate($request, [
             'lembar' => 'nullable|integer',
             'kode' => 'nullable|integer',
@@ -86,15 +86,164 @@ class SubkegLainSPPDController extends Controller
             'kepala_jabatan_di_tempat' => 'nullable|string|max:50',
             'nama_di_tempat' => 'nullable|string|max:50',
             'nip_di_tempat' => 'nullable|string|max:30',
+            'biaya_pergi' => 'nullable|numeric',
+            'biaya_pulang' => 'nullable|numeric',
+            'biaya_penginapan_4' => 'nullable|numeric',
+            'biaya_penginapan_3' => 'nullable|numeric',
+            'biaya_penginapan_2' => 'nullable|numeric',
+            'biaya_penginapan_1' => 'nullable|numeric',
+            'uang_harian' => 'nullable|numeric',
+            'keterangan_penerbangan' => 'nullable|string',
+            'keterangan_tol' => 'nullable|string',
+            'keterangan_lain' => 'nullable|string',
+            'biaya_penerbangan' => 'nullable|numeric',
+            'biaya_tol' => 'nullable|numeric',
+            'biaya_lain' => 'nullable|numeric',
         ]);
-        $sppd = SPPD::create($request->all());
+        $sppd = SPPD::create($request->except([
+            'biaya_pergi',
+            'biaya_pulang',
+            'biaya_penginapan_4',
+            'biaya_penginapan_3',
+            'biaya_penginapan_2',
+            'biaya_penginapan_1',
+            'uang_harian',
+            'keterangan_penerbangan',
+            'keterangan_tol',
+            'keterangan_lain',
+            'biaya_penerbangan',
+            'biaya_tol',
+            'biaya_lain'
+        ]));
+
+        $rincianBiaya = RincianBiayaSPPD::create([
+            'id_sppd' => $sppd->id,
+            'biaya_pergi' => $request->biaya_pergi,
+            'biaya_pulang' => $request->biaya_pulang,
+            'biaya_penginapan_4' => $request->biaya_penginapan_4,
+            'biaya_penginapan_3' => $request->biaya_penginapan_3,
+            'biaya_penginapan_2' => $request->biaya_penginapan_2,
+            'biaya_penginapan_1' => $request->biaya_penginapan_1,
+            'uang_harian' => $request->uang_harian,
+            'keterangan_penerbangan' => $request->keterangan_penerbangan,
+            'keterangan_tol' => $request->keterangan_tol,
+            'keterangan_lain' => $request->keterangan_lain,
+            'biaya_penerbangan' => $request->biaya_penerbangan,
+            'biaya_tol' => $request->biaya_tol,
+            'biaya_lain' => $request->biaya_lain,
+        ]);
+
+        // Perhitungan Biaya dan Lain-Lain
+        $pegawaiYangIkut = [
+            $rincianBiaya->sppd->pelaksana,
+            $rincianBiaya->sppd->pengikut_1,
+            $rincianBiaya->sppd->pengikut_2,
+            $rincianBiaya->sppd->pengikut_3,
+        ];
+
+        $totalPengikut = count($pegawaiYangIkut);
+        $totalGolIV = 0;
+        $totalGolIII = 0;
+        $totalGolII = 0;
+        $totalGolI = 0;
+
+        foreach ($pegawaiYangIkut as $p) {
+            if (!$p) continue;
+            $golongan = $p->golongan ?? null;
+            switch ($golongan) {
+                case 'I':
+                    $totalGolI++;
+                    break;
+                case 'III':
+                    $totalGolIII++;
+                    break;
+                case 'IV':
+                    $totalGolIV++;
+                    break;
+                case 'II':
+                case null:
+                case '':
+                    $totalGolII++;
+                    break;
+            }
+        }
+
+        // Gunakan preg_replace untuk menghapus semua karakter selain angka
+        $biayaPergi = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_pergi);
+        $biayaPulang = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_pulang);
+
+        $subtotalPergi = $totalPengikut * $biayaPergi;
+        $subtotalPulang = $totalPengikut * $biayaPulang;
+
+        $subtotalAngkutan = $subtotalPergi + $subtotalPulang;
+
+        // Terapkan hal yang sama untuk semua variabel biaya
+        $biayaPenginapan4 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_4);
+        $biayaPenginapan3 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_3);
+        $biayaPenginapan2 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_2);
+        $biayaPenginapan1 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_1);
+        $uangHarian = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->uang_harian);
+
+        $subtotalPenginapan4 = $totalGolIV * $biayaPenginapan4;
+        $subtotalPenginapan3 = $totalGolIII * $biayaPenginapan3;
+        $subtotalPenginapan2 = $totalGolII * $biayaPenginapan2;
+        $subtotalPenginapan1 = $totalGolI * $biayaPenginapan1;
+
+        $subtotalPenginapan = $subtotalPenginapan4 + $subtotalPenginapan3 + $subtotalPenginapan2 + $subtotalPenginapan1;
+
+        $subtotalHarian4 = $totalGolIV * $uangHarian;
+        $subtotalHarian3 = $totalGolIII * $uangHarian;
+        $subtotalHarian2 = $totalGolII * $uangHarian;
+        $subtotalHarian1 = $totalGolI * $uangHarian;
+
+        $subtotalHarian = $subtotalHarian4 + $subtotalHarian3 + $subtotalHarian2 + $subtotalHarian1;
+
+        $biayaPenerbangan = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penerbangan);
+        $biayaTol = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_tol);
+        $biayaLain = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_lain);
+
+        $subtotalLain = $biayaPenerbangan + $biayaTol + $biayaLain;
+
+        $totalBiaya = $subtotalAngkutan + $subtotalPenginapan + $subtotalHarian + $subtotalLain;
+
+        $biayaPelaksana = $uangHarian + $subtotalAngkutan + $subtotalLain;
+        $biayaPengikut = $uangHarian;
+
+        // Jangan lupa, saat menampilkan hasilnya di view, format kembali dengan titik
+        // Contoh di view Blade:
+        // <td>{{ number_format($subtotalAngkutan, 0, ',', '.') }}</td>
 
         // Nama file unik berdasarkan ID dan timestamp
         $fileName = 'sppd-' . $sppd->id . '-' . time() . '.pdf';
         $filePath = 'uploads/sppd/' . $fileName;
 
         // Menggunakan view blade sebagai template PDF
-        $pdf = PDF::loadView('admin.template.sppd', compact('sppd'));
+        $pdf = PDF::loadView('admin.template.sppd', compact(
+            'sppd',
+            'rincianBiaya',
+            'totalPengikut',
+            'totalGolIV',
+            'totalGolIII',
+            'totalGolII',
+            'totalGolI',
+            'subtotalPergi',
+            'subtotalPulang',
+            'subtotalAngkutan',
+            'subtotalPenginapan4',
+            'subtotalPenginapan3',
+            'subtotalPenginapan2',
+            'subtotalPenginapan1',
+            'subtotalPenginapan',
+            'subtotalHarian4',
+            'subtotalHarian3',
+            'subtotalHarian2',
+            'subtotalHarian1',
+            'subtotalHarian',
+            'subtotalLain',
+            'totalBiaya',
+            'biayaPelaksana',
+            'biayaPengikut',
+        ));
 
         // Simpan PDF ke direktori public/storage/uploads/sppd
         $pdf->save(public_path('storage/' . $filePath));
@@ -123,8 +272,9 @@ class SubkegLainSPPDController extends Controller
     {
         //
         $sppd = SPPD::findOrFail($id);
+        $rincianBiaya = RincianBiayaSPPD::find($id);
         $pegawai = Pegawai::all();
-        return view('admin.subkeg-lain.sppd.edit', compact('sppd', 'pegawai'));
+        return view('admin.subkeg-lain.sppd.edit', compact('sppd', 'pegawai', 'rincianBiaya'));
     }
 
     /**
@@ -132,7 +282,6 @@ class SubkegLainSPPDController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
         $this->validate($request, [
             'lembar' => 'nullable|integer',
             'kode' => 'nullable|integer',
@@ -159,19 +308,170 @@ class SubkegLainSPPDController extends Controller
             'kepala_jabatan_di_tempat' => 'nullable|string|max:50',
             'nama_di_tempat' => 'nullable|string|max:50',
             'nip_di_tempat' => 'nullable|string|max:30',
+            'biaya_pergi' => 'nullable|numeric',
+            'biaya_pulang' => 'nullable|numeric',
+            'biaya_penginapan_4' => 'nullable|numeric',
+            'biaya_penginapan_3' => 'nullable|numeric',
+            'biaya_penginapan_2' => 'nullable|numeric',
+            'biaya_penginapan_1' => 'nullable|numeric',
+            'uang_harian' => 'nullable|numeric',
+            'keterangan_penerbangan' => 'nullable|string',
+            'keterangan_tol' => 'nullable|string',
+            'keterangan_lain' => 'nullable|string',
+            'biaya_penerbangan' => 'nullable|numeric',
+            'biaya_tol' => 'nullable|numeric',
+            'biaya_lain' => 'nullable|numeric',
         ]);
+
         $sppd = SPPD::findOrFail($id);
+        $sppd->update($request->except([
+            'biaya_pergi',
+            'biaya_pulang',
+            'biaya_penginapan_4',
+            'biaya_penginapan_3',
+            'biaya_penginapan_2',
+            'biaya_penginapan_1',
+            'uang_harian',
+            'keterangan_penerbangan',
+            'keterangan_tol',
+            'keterangan_lain',
+            'biaya_penerbangan',
+            'biaya_tol',
+            'biaya_lain'
+        ]));
+        $rincianBiaya = RincianBiayaSPPD::findOrFail($id);
+        $rincianBiaya->update([
+            'id_sppd' => $sppd->id,
+            'biaya_pergi' => $request->biaya_pergi,
+            'biaya_pulang' => $request->biaya_pulang,
+            'biaya_penginapan_4' => $request->biaya_penginapan_4,
+            'biaya_penginapan_3' => $request->biaya_penginapan_3,
+            'biaya_penginapan_2' => $request->biaya_penginapan_2,
+            'biaya_penginapan_1' => $request->biaya_penginapan_1,
+            'uang_harian' => $request->uang_harian,
+            'keterangan_penerbangan' => $request->keterangan_penerbangan,
+            'keterangan_tol' => $request->keterangan_tol,
+            'keterangan_lain' => $request->keterangan_lain,
+            'biaya_penerbangan' => $request->biaya_penerbangan,
+            'biaya_tol' => $request->biaya_tol,
+            'biaya_lain' => $request->biaya_lain,
+        ]);
 
-        unlink(public_path('storage/uploads/sppd/' . $sppd->file_surat));
+        // Perhitungan Biaya dan Lain-Lain
+        $pegawaiYangIkut = [
+            $rincianBiaya->sppd->pelaksana,
+            $rincianBiaya->sppd->pengikut_1,
+            $rincianBiaya->sppd->pengikut_2,
+            $rincianBiaya->sppd->pengikut_3,
+        ];
 
-        $sppd->update($request->all());
+        $totalPengikut = count($pegawaiYangIkut);
+        $totalGolIV = 0;
+        $totalGolIII = 0;
+        $totalGolII = 0;
+        $totalGolI = 0;
+
+        foreach ($pegawaiYangIkut as $p) {
+            if (!$p) continue;
+            $golongan = $p->golongan ?? null;
+            switch ($golongan) {
+                case 'I':
+                    $totalGolI++;
+                    break;
+                case 'III':
+                    $totalGolIII++;
+                    break;
+                case 'IV':
+                    $totalGolIV++;
+                    break;
+                case 'II':
+                case null:
+                case '':
+                    $totalGolII++;
+                    break;
+            }
+        }
+
+        // Gunakan preg_replace untuk menghapus semua karakter selain angka
+        $biayaPergi = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_pergi);
+        $biayaPulang = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_pulang);
+
+        $subtotalPergi = $totalPengikut * $biayaPergi;
+        $subtotalPulang = $totalPengikut * $biayaPulang;
+
+        $subtotalAngkutan = $subtotalPergi + $subtotalPulang;
+
+        // Terapkan hal yang sama untuk semua variabel biaya
+        $biayaPenginapan4 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_4);
+        $biayaPenginapan3 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_3);
+        $biayaPenginapan2 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_2);
+        $biayaPenginapan1 = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penginapan_1);
+        $uangHarian = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->uang_harian);
+
+        $subtotalPenginapan4 = $totalGolIV * $biayaPenginapan4;
+        $subtotalPenginapan3 = $totalGolIII * $biayaPenginapan3;
+        $subtotalPenginapan2 = $totalGolII * $biayaPenginapan2;
+        $subtotalPenginapan1 = $totalGolI * $biayaPenginapan1;
+
+        $subtotalPenginapan = $subtotalPenginapan4 + $subtotalPenginapan3 + $subtotalPenginapan2 + $subtotalPenginapan1;
+
+        $subtotalHarian4 = $totalGolIV * $uangHarian;
+        $subtotalHarian3 = $totalGolIII * $uangHarian;
+        $subtotalHarian2 = $totalGolII * $uangHarian;
+        $subtotalHarian1 = $totalGolI * $uangHarian;
+
+        $subtotalHarian = $subtotalHarian4 + $subtotalHarian3 + $subtotalHarian2 + $subtotalHarian1;
+
+        $biayaPenerbangan = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_penerbangan);
+        $biayaTol = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_tol);
+        $biayaLain = (int) preg_replace('/[^0-9]/', '', $rincianBiaya->biaya_lain);
+
+        $subtotalLain = $biayaPenerbangan + $biayaTol + $biayaLain;
+
+        $totalBiaya = $subtotalAngkutan + $subtotalPenginapan + $subtotalHarian + $subtotalLain;
+
+        $biayaPelaksana = $uangHarian + $subtotalAngkutan + $subtotalLain;
+        $biayaPengikut = $uangHarian;
+
+        // Jangan lupa, saat menampilkan hasilnya di view, format kembali dengan titik
+        // Contoh di view Blade:
+        // <td>{{ number_format($subtotalAngkutan, 0, ',', '.') }}</td>
+
+        if ($sppd->file_surat) {
+            unlink(public_path('storage/uploads/sppd/' . $sppd->file_surat));
+        }
 
         // Nama file unik berdasarkan ID dan timestamp
         $fileName = 'sppd-' . $sppd->id . '-' . time() . '.pdf';
         $filePath = 'uploads/sppd/' . $fileName;
 
         // Menggunakan view blade sebagai template PDF
-        $pdf = PDF::loadView('admin.template.sppd', compact('sppd'));
+        $pdf = PDF::loadView('admin.template.sppd', compact(
+            'sppd',
+            'rincianBiaya',
+            'totalPengikut',
+            'totalGolIV',
+            'totalGolIII',
+            'totalGolII',
+            'totalGolI',
+            'subtotalPergi',
+            'subtotalPulang',
+            'subtotalAngkutan',
+            'subtotalPenginapan4',
+            'subtotalPenginapan3',
+            'subtotalPenginapan2',
+            'subtotalPenginapan1',
+            'subtotalPenginapan',
+            'subtotalHarian4',
+            'subtotalHarian3',
+            'subtotalHarian2',
+            'subtotalHarian1',
+            'subtotalHarian',
+            'subtotalLain',
+            'totalBiaya',
+            'biayaPelaksana',
+            'biayaPengikut',
+        ));
 
         // Simpan PDF ke direktori public/storage/uploads/sppd
         $pdf->save(public_path('storage/' . $filePath));
@@ -190,7 +490,9 @@ class SubkegLainSPPDController extends Controller
     {
         //
         $sppd = SPPD::find($id);
-        unlink(public_path('storage/uploads/sppd/' . $sppd->file_surat));
+        if ($sppd->file_surat) {
+            unlink(public_path('storage/uploads/sppd/' . $sppd->file_surat));
+        }
         $sppd->delete();
         return redirect()->route('subkeg-lain.sppd.index')->with('success', 'Surat Perjalanan Dinas berhasil dihapus.');
     }
